+++ src/frontend/components/StudentLiveWrapper/index.spec.tsx
@@ -1,8 +1,10 @@
-import React from 'react';
-import fetchMock from 'fetch-mock';
 import { act, render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+import faker from 'faker';
+import fetchMock from 'fetch-mock';
+import React from 'react';
 
+import { pushAttendance } from 'data/sideEffects/pushAttendance';
 import { useChatItemState } from 'data/stores/useChatItemsStore';
 import {
   LivePanelItem,
@@ -10,16 +12,19 @@ import {
 } from 'data/stores/useLivePanelState';
 import { useLiveStateStarted } from 'data/stores/useLiveStateStarted';
 import { useParticipantWorkflow } from 'data/stores/useParticipantWorkflow';
+import { PictureInPictureProvider } from 'data/stores/usePictureInPicture/index';
+import { createPlayer } from 'Player/createPlayer';
 import { LiveModeType, liveState } from 'types/tracks';
 import { PersistentStore } from 'types/XMPP';
-import { videoMockFactory } from 'utils/tests/factories';
+import { getOrInitAnonymousId } from 'utils/getOrInitAnonymousId';
+import {
+  sharedLiveMediaMockFactory,
+  videoMockFactory,
+} from 'utils/tests/factories';
 import { wrapInIntlProvider } from 'utils/tests/intl';
-import { createPlayer } from 'Player/createPlayer';
+import { wrapInRouter } from 'utils/tests/router';
 
 import { StudentLiveWrapper } from '.';
-
-import { pushAttendance } from 'data/sideEffects/pushAttendance';
-import { getOrInitAnonymousId } from 'utils/getOrInitAnonymousId';
 
 const mockVideo = videoMockFactory();
 jest.mock('data/appData', () => ({
@@ -146,7 +151,9 @@ describe('<StudentLiveWrapper /> as a viewer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -209,7 +216,9 @@ describe('<StudentLiveWrapper /> as a viewer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -270,7 +279,9 @@ describe('<StudentLiveWrapper /> as a viewer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -331,7 +342,9 @@ describe('<StudentLiveWrapper /> as a viewer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -386,7 +399,9 @@ describe('<StudentLiveWrapper /> as a viewer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -422,7 +437,9 @@ describe('<StudentLiveWrapper /> as a viewer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -439,6 +456,66 @@ describe('<StudentLiveWrapper /> as a viewer', () => {
     );
 
     screen.getByText('No title');
+  });
+
+  it('displays the video in the picture and shared media in background when a media is shared', async () => {
+    const videoId = faker.datatype.uuid();
+    const sharedLiveMedia = sharedLiveMediaMockFactory({ video: videoId });
+    const video = videoMockFactory({
+      id: videoId,
+      active_shared_live_media: sharedLiveMedia,
+      active_shared_live_media_page: 1,
+      title: 'live title',
+      live_info: {},
+      live_state: liveState.RUNNING,
+      live_type: LiveModeType.RAW,
+      shared_live_medias: [sharedLiveMedia],
+      xmpp: {
+        bosh_url: 'https://xmpp-server.com/http-bind',
+        converse_persistent_store: PersistentStore.LOCALSTORAGE,
+        websocket_url: null,
+        conference_url:
+          '870c467b-d66e-4949-8ee5-fcf460c72e88@conference.xmpp-server.com',
+        prebind_url: 'https://xmpp-server.com/http-pre-bind',
+        jid: 'xmpp-server.com',
+      },
+    });
+    useChatItemState.setState({
+      hasReceivedMessageHistory: true,
+    });
+
+    const { container } = render(
+      wrapInIntlProvider(
+        wrapInRouter(
+          <PictureInPictureProvider value={{ reversed: true }}>
+            <StudentLiveWrapper video={video} playerType={'player_type'} />
+          </PictureInPictureProvider>,
+        ),
+      ),
+    );
+
+    await waitFor(() =>
+      // The player is created
+      expect(mockCreatePlayer).toHaveBeenCalledWith(
+        'player_type',
+        expect.any(Element),
+        expect.anything(),
+        video,
+        'en',
+        expect.any(Function),
+      ),
+    );
+
+    const pipMaster = container.querySelector('#picture-in-picture-master');
+    const pipSlave = container.querySelector('#picture-in-picture-slave');
+    expect(pipMaster!.getElementsByTagName('source')[0]).toHaveAttribute(
+      'src',
+      `https://example.com/mp4/1080`,
+    );
+    expect(pipSlave?.getElementsByTagName('img')[0]).toHaveAttribute(
+      'src',
+      `https://example.com/sharedLiveMedia/${sharedLiveMedia.id}/1`,
+    );
   });
 });
 
@@ -502,7 +579,9 @@ describe('<StudentLiveWrapper /> as a streamer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -556,7 +635,9 @@ describe('<StudentLiveWrapper /> as a streamer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -608,7 +689,9 @@ describe('<StudentLiveWrapper /> as a streamer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -659,7 +742,9 @@ describe('<StudentLiveWrapper /> as a streamer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -705,7 +790,9 @@ describe('<StudentLiveWrapper /> as a streamer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -756,7 +843,9 @@ describe('<StudentLiveWrapper /> as a streamer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -797,7 +886,9 @@ describe('<StudentLiveWrapper /> as a streamer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -855,7 +946,9 @@ describe('<StudentLiveWrapper /> as a streamer', () => {
 
     render(
       wrapInIntlProvider(
-        <StudentLiveWrapper video={video} playerType={'player_type'} />,
+        <PictureInPictureProvider value={{ reversed: true }}>
+          <StudentLiveWrapper video={video} playerType={'player_type'} />
+        </PictureInPictureProvider>,
       ),
     );
 
@@ -893,6 +986,60 @@ describe('<StudentLiveWrapper /> as a streamer', () => {
         'en',
         expect.any(Function),
       ),
+    );
+  });
+
+  it('displays the jitsi player in the picture and shared media in background when a media is shared', async () => {
+    const videoId = faker.datatype.uuid();
+    const sharedLiveMedia = sharedLiveMediaMockFactory({ video: videoId });
+    const video = videoMockFactory({
+      id: videoId,
+      active_shared_live_media: sharedLiveMedia,
+      active_shared_live_media_page: 1,
+      title: 'live title',
+      live_info: {
+        jitsi: {
+          domain: 'meet.jit.si',
+          external_api_url: 'https://meet.jit.si/external_api.js',
+          config_overwrite: {},
+          interface_config_overwrite: {},
+          room_name: 'jitsi_conference',
+        },
+      },
+      live_state: liveState.IDLE,
+      live_type: LiveModeType.JITSI,
+      shared_live_medias: [sharedLiveMedia],
+      xmpp: {
+        bosh_url: 'https://xmpp-server.com/http-bind',
+        converse_persistent_store: PersistentStore.LOCALSTORAGE,
+        websocket_url: null,
+        conference_url:
+          '870c467b-d66e-4949-8ee5-fcf460c72e88@conference.xmpp-server.com',
+        prebind_url: 'https://xmpp-server.com/http-pre-bind',
+        jid: 'xmpp-server.com',
+      },
+    });
+    useChatItemState.setState({
+      hasReceivedMessageHistory: true,
+    });
+
+    const { container } = render(
+      wrapInIntlProvider(
+        wrapInRouter(
+          <PictureInPictureProvider value={{ reversed: true }}>
+            <StudentLiveWrapper video={video} playerType={'player_type'} />
+          </PictureInPictureProvider>,
+        ),
+      ),
+    );
+
+    expect(mockJitsi).toHaveBeenCalled();
+
+    container.querySelector('#picture-in-picture-master');
+    const pipSlave = container.querySelector('#picture-in-picture-slave');
+    expect(pipSlave!.getElementsByTagName('img')[0]).toHaveAttribute(
+      'src',
+      `https://example.com/sharedLiveMedia/${sharedLiveMedia.id}/1`,
     );
   });
 });